refactor(omegaform): extract createMeta + AST helpers into meta/createMeta
